refactor : 공통 모듈 패키지 정리에 따른 경로 수정
