+++ src/main/java/com/dss/realworld/comment/app/CommentServiceImpl.java
@@ -8,9 +8,9 @@ import com.dss.realworld.comment.api.dto.CommentDto;
 import com.dss.realworld.comment.domain.Comment;
 import com.dss.realworld.comment.domain.repository.CommentRepository;
 import com.dss.realworld.common.dto.AuthorDto;
-import com.dss.realworld.error.exception.ArticleNotFoundException;
-import com.dss.realworld.error.exception.CommentNotFoundException;
-import com.dss.realworld.error.exception.UserNotFoundException;
+import com.dss.realworld.common.error.exception.ArticleNotFoundException;
+import com.dss.realworld.common.error.exception.CommentNotFoundException;
+import com.dss.realworld.common.error.exception.UserNotFoundException;
 import com.dss.realworld.user.domain.User;
 import com.dss.realworld.user.domain.repository.FollowingRepository;
 import com.dss.realworld.user.domain.repository.UserRepository;
