+++ src/test/java/com/dss/realworld/article/app/ArticleServiceTest.java
@@ -10,8 +10,8 @@ import com.dss.realworld.article.domain.repository.ArticleTagRepository;
 import com.dss.realworld.article.domain.repository.ArticleUsersRepository;
 import com.dss.realworld.comment.domain.Comment;
 import com.dss.realworld.comment.domain.repository.CommentRepository;
-import com.dss.realworld.error.exception.ArticleNotFoundException;
-import com.dss.realworld.error.exception.DuplicateFavoriteException;
+import com.dss.realworld.common.error.exception.ArticleNotFoundException;
+import com.dss.realworld.common.error.exception.DuplicateFavoriteException;
 import com.dss.realworld.user.domain.Following;
 import com.dss.realworld.user.domain.User;
 import com.dss.realworld.user.domain.repository.FollowingRepository;
